refactoring, ts error fix
